Calculate the Ecosystem Service for daily fish catch using Method 2 and monthly thresholds across multiple CFi's

Calculate the Ecosystem Service for daily fish catch using Method 2 and monthly thresholds across multiple CFi's

ES_DFC_M(
  x,
  Jan_t,
  Feb_t,
  Mar_t,
  Apr_t,
  May_t,
  Jun_t,
  Jul_t,
  Aug_t,
  Sep_t,
  Oct_t,
  Nov_t,
  Dec_t,
  S_Date,
  E_Date
)

Arguments

x

input file

Jan_t

Daily catch threshold weight for January

Feb_t

Daily catch threshold weight for February

Mar_t

Daily catch threshold weight for March

Apr_t

Daily catch threshold weight for April

May_t

Daily catch threshold weight for May

Jun_t

Daily catch threshold weight for June

Jul_t

Daily catch threshold weight for July

Aug_t

Daily catch threshold weight for August

Sep_t

Daily catch threshold weight for September

Oct_t

Daily catch threshold weight for October

Nov_t

Daily catch threshold weight for November

Dec_t

Daily catch threshold weight for December

S_Date

Start date for assessment period

E_Date

End date for assessment period

Examples

ES_DFC_M(DailyCatch, 50, 50, 20, 20, 10, 50, 50, 50, 50, 50, 50, 50, '2021-01-01', '2021-12-31')
#>      Parameter  F1       F2       F3 ESI_F1_3
#> 1 WeightCaught 100 98.46293 84.80139 7.912328